requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
pydantic[email]>=2.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9
python-dotenv>=1.0.1
httpx>=0.27.0
ccxt>=4.3.20
apscheduler>=3.10.4
websockets>=12.0
psutil>=5.9.0
# ── LangChain + LangGraph (AI-powered code analysis) ──────────────────────────
langchain>=0.3.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# ── Build / packaging ────────────────────────────────────────────────────────
# Installed automatically by build-exes.bat if missing. Pinning here keeps
# everyone's frozen exes building from the same PyInstaller version.
pyinstaller>=6.10.0
